Add catch-all route if one not regeisterd yet (#8)
* Add root handling to allow middleware on unregistered paths

This update introduces root handling functionality to the group routing. Now, if no root route ("/") is registered, a not found handler is automatically registered for the root route. This helps in utilizing middleware for all routes such as logging. Additionally, the middleware wrapping process has been refactored into a separate function to streamline the code and improve readability. The update also includes related tests and minor tweaks to existing tests. #7

* update docs

* Implement option to disable automatic NotFoundHandler registration

* add test for DisableNotFoundHandler beheaivour