Add support for "path-less" patterns
This add support for patterns without path. When using the pattern {$} the
path will be set to an empty string. This allows to handle a mounted group
root url without trailing slash.

For example:
     	router := routegroup.New(mux)
     	usersGroup := router.Mount("/api/v1/users")
        usersGroup.Use(middleware1, middleware2)
     	usersGroup.Handle("{$}", handler) // this will handle /api/v1/users
     	usersGroup.Handle("/new", handler) // this will handle /api/v1/users/new

Without this, you need to regsiter the "root" handler on the parent
group and explicitly add middlewares/... to the path. Like this:

     	router := routegroup.New(mux)
     	router.With(middleware1, middleware2).Handle("/api/v1/users", handler)
     	usersGroup := router.Mount("/api/v1/users")
        usersGroup.Use(middleware1, middleware2)
     	usersGroup.Handle("/new", handler) // this will handle /api/v1/users/new

Lowering readability and maintainability.